Split data size and GPU alloc part 1

+++ src/kernel.h
@@ -5,6 +5,6 @@ public:
 	Kernel() = default;
 	virtual ~Kernel() = default;
 
-	virtual void runCpu(int workItemId, int workGroupSize) = 0;
-	virtual void runGpu(int deviceId, int workItemId, int workGroupSize) = 0;
+	virtual void runCpu(uint64_t workItemId, uint64_t workGroupSize) = 0;
+	virtual void runGpu(uint64_t deviceId, uint64_t workItemId, uint64_t workGroupSize) = 0;
 };
